WIP: switch to using workers
More debugging needed before this works but the basics are in place.
